+++ CHANGELOG.md
@@ -5,12 +5,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Support for INT8 and UINT8 channels (used by vacuum valves for VTUX)
+- Support for AP modules with different channel types
+- System tests for VTUX-AS
+- Information on channels in system_information docu
+
 ### Changed
 - Removed optional parameter `raw` from `read_module_parameter()`. Added `read_module_parameter_enum_str()` instead.
 - Removed optional parameter `outputs_only` from `read_channel(s)()`. Added `read_output_channel(s)()` instead.
 - Updated examples
-# Added
+
+### Fixed
+- Updated ModbusException handling for ConnectionAbortedError and corresponding unittest
+- Corrected (U)INT16 handling and corresponding unittests
 - System tests for VTUX-AS
+- Typo in CPX-AP docu (import)
 
 ## v0.5.3 - 09.08.24
 ### Added